eth_pubsub support (#103)
* Pubsub service setup

* Add LatestHeader StorageValue, change Module name

* Implement Kind::NewHeads for eth_subscribe

* Styling fix

* Add .cargo-remote.toml to .gitignore

* Update Cargo.lock

* pallet_ethereum LatestHeader to LatestBlock

* Refactor to accommodate new Kinds

* Implement Kind::Logs

* Move Vec<Log> build to its own function

* Added missing fields from Log response

* Implement Kind::NewPendingTransactions

* Implement eth_unsubscribe

* Implement filtering for Kind::Logs

* Small fixes

* Add stream_build macro

* Adapt to frontier consensus

* Indentation cleanup

* Formatting and cleanup

* Remove unused imports

* Remove Pending.take and add on_initialize

* Fix Kind:NewPendingTransactions

* Add NetworkService to pubsub rpc handler

* Implement Kind::Syncing

* Fix checker

* pallet_ethereum handle Log response from evm execution

* Runtime update substrate#12d0b7e

* Return Vec<Log> in handle_exec

* Fix checker

* Runtime update substrate#93b2c36

* Bump jsonrpc-ore and jsonrpc-pubsub version

* Update Cargo.lock

* Update pallet-ethereum mock

* Add missing const

* Move internal error handler to crate root

* Update pubsub to one-to-many

* Skip previous block hash as parent test
