+++ frame/ethereum/src/lib.rs
@@ -61,7 +61,7 @@ pub trait Trait: frame_system::Trait<Hash=H256> + pallet_balances::Trait + palle
 }
 
 decl_storage! {
-	trait Store for Module<T: Trait> as Example {
+	trait Store for Module<T: Trait> as Ethereum {
 		/// Current building block's transactions and receipts.
 		Pending: Vec<(ethereum::Transaction, TransactionStatus, ethereum::Receipt)>;
 
@@ -162,6 +162,11 @@ decl_module! {
 		fn on_finalize(n: T::BlockNumber) {
 			<Module<T>>::store_block();
 		}
+
+		fn on_initialize(n: T::BlockNumber) -> frame_support::weights::Weight {
+			Pending::kill();
+			0
+		}
 	}
 }
 
@@ -228,12 +233,10 @@ impl<T: Trait> Module<T> {
 	}
 
 	fn store_block() {
-		let pending = Pending::take();
-
 		let mut transactions = Vec::new();
 		let mut statuses = Vec::new();
 		let mut receipts = Vec::new();
-		for (transaction, status, receipt) in pending {
+		for (transaction, status, receipt) in Pending::get() {
 			transactions.push(transaction);
 			statuses.push(status);
 			receipts.push(receipt);
@@ -272,7 +275,7 @@ impl<T: Trait> Module<T> {
 		let hash = Self::ethereum_block_hash(&header);
 
 		let block = ethereum::Block {
-			header,
+			header: header.clone(),
 			transactions: transactions.clone(),
 			ommers,
 		};
@@ -337,7 +340,7 @@ impl<T: Trait> Module<T> {
 
 		let (status, gas_used) = match transaction.action {
 			ethereum::TransactionAction::Call(target) => {
-				let (_, _, gas_used) = Self::handle_exec(
+				let (_, _, gas_used, logs) = Self::handle_exec(
 					pallet_evm::Module::<T>::execute_call(
 						source,
 						target,
@@ -356,12 +359,21 @@ impl<T: Trait> Module<T> {
 					from: source,
 					to: Some(target),
 					contract_address: None,
-					logs: Vec::new(), // TODO: feed in logs.
+					logs: {
+						logs.into_iter()
+						.map(|log| {
+							Log {
+								address: log.address,
+								topics: log.topics,
+								data: log.data
+							}
+						}).collect()
+					},
 					logs_bloom: Bloom::default(), // TODO: feed in bloom.
 				}, gas_used)
 			},
 			ethereum::TransactionAction::Create => {
-				let (_, contract_address, gas_used) = Self::handle_exec(
+				let (_, contract_address, gas_used, logs) = Self::handle_exec(
 					pallet_evm::Module::<T>::execute_create(
 						source,
 						transaction.input.clone(),
@@ -379,7 +391,16 @@ impl<T: Trait> Module<T> {
 					from: source,
 					to: None,
 					contract_address: Some(contract_address),
-					logs: Vec::new(), // TODO: feed in logs.
+					logs: {
+						logs.into_iter()
+						.map(|log| {
+							Log {
+								address: log.address,
+								topics: log.topics,
+								data: log.data
+							}
+						}).collect()
+					},
 					logs_bloom: Bloom::default(), // TODO: feed in bloom.
 				}, gas_used)
 			},
@@ -389,7 +410,7 @@ impl<T: Trait> Module<T> {
 			state_root: H256::default(), // TODO: should be okay / error status.
 			used_gas: gas_used,
 			logs_bloom: Bloom::default(), // TODO: set this.
-			logs: Vec::new(), // TODO: set this.
+			logs: status.clone().logs,
 		};
 
 		Pending::append((transaction, status, receipt));
@@ -397,7 +418,8 @@ impl<T: Trait> Module<T> {
 		Ok(())
 	}
 
-	fn handle_exec<R>(res: (ExitReason, R, U256)) -> Result<(ExitReason, R, U256), Error<T>> {
+	fn handle_exec<R>(res: (ExitReason, R, U256, Vec<pallet_evm::Log>))
+		-> Result<(ExitReason, R, U256, Vec<pallet_evm::Log>), Error<T>> {
 		match res.0 {
 			ExitReason::Succeed(_s) => Ok(res),
 			ExitReason::Error(e) => Err(Self::parse_exit_error(e)),
