+++ rpc/core/Cargo.toml
@@ -7,10 +7,10 @@ description = "RPC traits of Ethereum."
 license = "GPL-3.0"
 
 [dependencies]
-jsonrpc-core = "14.0.3"
+jsonrpc-core = "15.0.0"
 jsonrpc-core-client = "14.0.3"
 jsonrpc-derive = "14.0.3"
-jsonrpc-pubsub = "14.0.5"
+jsonrpc-pubsub = "15.0.0"
 rustc-hex = "2.1.0"
 ethereum-types = "0.9.0"
 serde = { version = "1.0", features = ["derive"] }
