+++ rpc/core/src/lib.rs
@@ -23,7 +23,7 @@ mod net;
 mod web3;
 
 pub use eth::{EthApi, EthApiServer, EthFilterApi};
-pub use eth_pubsub::EthPubSubApi;
+pub use eth_pubsub::{EthPubSubApi, EthPubSubApiServer};
 pub use eth_signing::EthSigningApi;
 pub use net::{NetApi, NetApiServer};
 pub use web3::Web3Api;
