+++ rpc/core/src/types/mod.rs
@@ -38,7 +38,7 @@ pub use self::bytes::Bytes;
 pub use self::block::{RichBlock, Block, BlockTransactions, Header, RichHeader, Rich};
 pub use self::block_number::BlockNumber;
 pub use self::call_request::CallRequest;
-pub use self::filter::{Filter, FilterChanges, VariadicValue};
+pub use self::filter::{Filter, FilterChanges, VariadicValue, FilterAddress, Topic};
 pub use self::index::Index;
 pub use self::log::Log;
 pub use self::receipt::Receipt;
