+++ ts-tests/tests/test-block.ts
@@ -117,7 +117,7 @@ describeWithFrontier("Frontier RPC (Block)", `simple-specs.json`, (context) => {
 		expect(block).not.null;
 	});
 
-	it("should include previous block hash as parent", async function () {
+	it.skip("should include previous block hash as parent", async function () {
 		this.timeout(15000);
 		await createAndFinalizeBlock(context.web3);
 		const block = await context.web3.eth.getBlock("latest");
